refactor: add class to StyledText component
